requirements.txt
requests
beautifulsoup4
python-telegram-bot==20.3
selenium
webdriver-manager
flask
